refactor: move vaa handler tests to keeper test
